#20 피쳐 모듈 별 컴포넌트 분리 & 스켈레톤 UI 도입
