+++ core/designsystem/src/main/java/com/example/designsystem/theme/Color.kt
@@ -13,3 +13,4 @@ val Pink40 = Color(0xFF7D5260)
 val Black90 = Color(0xFF222222)
 val Gray70 = Color(0xFF7A7A7A)
 val Gray30 = Color(0xFFEAEBEF)
+val Gray20 = Color(0xFFF5F6FA)
