+++ feature/detail/build.gradle.kts
@@ -70,5 +70,4 @@ dependencies {
     implementation(libs.androidx.navigation.compose)
 
     implementation(libs.coil.compose)
-    implementation(libs.accompanist.systemuicontroller)
 }
