+++ feature/detail/src/main/java/com/example/detail/component/DetailBottom.kt
@@ -1,0 +1,60 @@
+package com.example.detail.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.designsystem.R
+
+@Composable
+fun DetailBottom(
+    title: String?,
+    description: String?
+) {
+    Column(
+        modifier = Modifier.padding(
+            start = 12.dp,
+            end = 12.dp,
+            bottom = 12.dp
+        ),
+        verticalArrangement = Arrangement.spacedBy(10.dp)
+    ) {
+        Text(
+            text = title ?: "Title",
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+            style = TextStyle(
+                fontSize = 20.sp,
+                fontFamily = FontFamily(Font(R.font.pretendard_semi_bold))
+            ),
+            color = Color.White
+        )
+        Text(
+            text = description ?: "description\ndescription은 최대 2줄",
+            maxLines = 2,
+            overflow = TextOverflow.Ellipsis,
+            style = TextStyle(
+                fontSize = 16.sp,
+                fontFamily = FontFamily(Font(R.font.pretendard_medium))
+            ),
+            color = Color.White
+        )
+        Text(
+            text = "#tag #tag #tag #tag",
+            style = TextStyle(
+                fontSize = 16.sp,
+                fontFamily = FontFamily(Font(R.font.pretendard_medium))
+            ),
+            color = Color.White
+        )
+    }
+}
